Consolidate all graphQL mock responses to their respective mock files
