+++ .tool-versions
@@ -1,0 +1,4 @@
+erlang 22.2.6
+ruby 2.7.1
+nodejs 13.8.0
+elixir 1.9.3-otp-22
