+++ assets/js/components/Collection/Collection.test.jsx
@@ -1,23 +1,10 @@
 import React from "react";
-import { Route } from "react-router-dom";
 import Collection from "./Collection";
 import { renderWithRouterApollo } from "../../services/testing-helpers";
-
-const mockCollection = {
-  adminEmail: "test@test.com",
-  description: "Test arrays keyword arrays arrays arrays arrays",
-  featured: false,
-  findingAidUrl: "http://go.com",
-  id: "7a6c7b35-41a6-465a-9be2-0587c6b39ae0",
-  keywords: ["yo", "foo", "bar", "dude", "hey"],
-  name: "Ima collection",
-  published: false,
-  works: [],
-  representativeWork: null,
-};
+import { collectionMock } from "./collection.gql.mock";
 
 function setUpTests() {
-  return renderWithRouterApollo(<Collection collection={mockCollection} />, {});
+  return renderWithRouterApollo(<Collection collection={collectionMock} />, {});
 }
 
 it("renders Collection component", async () => {
@@ -28,8 +15,6 @@ it("renders Collection component", async () => {
 
 it("renders collection properties", async () => {
   const { getByText } = setUpTests();
-  expect(getByText("test@test.com")).toBeInTheDocument();
-  expect(
-    getByText("Test arrays keyword arrays arrays arrays arrays")
-  ).toBeInTheDocument();
+  expect(getByText("admin@nu.com")).toBeInTheDocument();
+  expect(getByText("Collection description lorem ipsum")).toBeInTheDocument();
 });
