+++ assets/js/components/Collection/ListRow.test.jsx
@@ -1,18 +1,12 @@
 import React from "react";
 import CollectionListRow from "./ListRow";
 import { renderWithRouter } from "../../services/testing-helpers";
-
-const collection = {
-  description: "collection description",
-  id: "01DVRVWZD0FJ3CJMQZC9S9HSTF",
-  keywords: ["test, testing, tested"],
-  name: "Ima Collection"
-};
+import { collectionMock } from "./collection.gql.mock";
 
 describe("CollectionListRow component", () => {
   it("renders the root element", () => {
     const { getByTestId } = renderWithRouter(
-      <CollectionListRow collection={collection} />
+      <CollectionListRow collection={collectionMock} />
     );
     expect(getByTestId("collection-list-row")).toBeInTheDocument();
   });
