+++ assets/js/components/Collection/Search.test.jsx
@@ -1,46 +1,10 @@
 import React from "react";
 import CollectionSearch from "./Search";
 import { renderWithRouter } from "../../services/testing-helpers";
+import { collectionMock } from "./collection.gql.mock";
 
-const mockCollection = {
-  adminEmail: "admin@nu.com",
-  description: "asdf asdfasdf",
-  representativeImage: "https://thisIsTest.com",
-  featured: true,
-  findingAidUrl: "http://something.com",
-  id: "01DWHQQYTVKC2THHW8SHRBH2XP",
-  keywords: ["any", " work", "foo", "bar"],
-  name: "Great collection",
-  published: false,
-  works: [
-    {
-      id: "1id-23343432",
-      accessionNumber: "accessNumber1",
-      representativeImage: "repImage1url.com",
-      descriptiveMetadata: {
-        title: "Title 1",
-      },
-    },
-    {
-      id: "2is-234o24332-id",
-      accessionNumber: "accessNumber2",
-      representativeImage: "repImage2url.com",
-      descriptiveMetadata: {
-        title: null,
-      },
-    },
-    {
-      id: "3id-23sd7343432",
-      accessionNumber: "accessNumber3",
-      representativeImage: null,
-      descriptiveMetadata: {
-        title: "Test title goes here",
-      },
-    },
-  ],
-};
 function setupTests() {
-  return renderWithRouter(<CollectionSearch collection={mockCollection} />);
+  return renderWithRouter(<CollectionSearch collection={collectionMock} />);
 }
 describe("CollectionSearch component", () => {
   it("renders the root element", () => {
@@ -50,7 +14,7 @@ describe("CollectionSearch component", () => {
 
   it("displays number of works", () => {
     const { getByTestId, debug } = setupTests();
-    expect(getByTestId("number-of-works").innerHTML).toBe("3 results...");
+    expect(getByTestId("number-of-works").innerHTML).toBe("2 results...");
   });
 
   it("displays work title", () => {
@@ -75,7 +39,7 @@ describe("CollectionSearch component", () => {
 
   it("displays default work image", () => {
     const { getByTestId } = setupTests();
-    const el = getByTestId("work-image-3id-23sd7343432");
+    const el = getByTestId("work-image-2is-234o24332-id");
     expect(el.getAttribute("src")).toEqual("/images/480x480.png");
   });
 });
