+++ assets/js/components/IngestSheet/List.test.jsx
@@ -2,7 +2,6 @@ import React from "react";
 import { render, cleanup } from "@testing-library/react";
 import IngestSheetList from "./List";
 import "@testing-library/jest-dom/extend-expect";
-import { IngestSheets } from "../../mock-data/ingestSheets";
 import { renderWithRouter } from "../../services/testing-helpers";
 
 const projectId = "abcdefg123";
