+++ assets/js/components/Work/ListItem.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import WorkListItem from "./ListItem";
-import { mockWork as work } from "../../services/testing-helpers";
+import { mockWork as work } from "./work.gql.mock";
 import { renderWithRouter } from "../../services/testing-helpers";
 
 const workObject = {
@@ -42,12 +42,12 @@ describe("Shows Work content", () => {
   it("Displays Accession Number", () => {
     const { getByTestId } = setupTests();
     expect(getByTestId("result-item-accession-number").innerHTML).toBe(
-      "Example-34"
+      "Donohue_001"
     );
   });
   it("Displays FileSets Length", () => {
     const { getByTestId } = setupTests();
-    expect(getByTestId("result-item-filesets-length").innerHTML).toBe("2");
+    expect(getByTestId("result-item-filesets-length").innerHTML).toBe("4");
   });
   it("Displays Updated Date", () => {
     const { getByTestId } = setupTests();
