+++ assets/js/components/Work/Work.test.js
@@ -1,10 +1,8 @@
 import React from "react";
 import Work from "./Work";
 import { waitFor } from "@testing-library/react";
-import {
-  renderWithRouterApollo,
-  mockWork,
-} from "../../services/testing-helpers";
+import { renderWithRouterApollo } from "../../services/testing-helpers";
+import { mockWork } from "./work.gql.mock";
 import { IIIF_SERVER_URL } from "../IIIF/iiif.gql";
 import {
   codeListAuthorityMock,
