+++ assets/js/screens/Collection/List.test.js
@@ -1,35 +1,10 @@
 import React from "react";
 import ScreensCollectionList from "./List";
-import { GET_COLLECTIONS } from "../../components/Collection/collection.gql.js";
 import { renderWithRouterApollo } from "../../services/testing-helpers";
 import { Route } from "react-router-dom";
 import { waitFor } from "@testing-library/react";
-
-const mocks = [
-  {
-    request: {
-      query: GET_COLLECTIONS,
-    },
-    result: {
-      data: {
-        collections: [
-          {
-            adminEmail: null,
-            description: null,
-            featured: false,
-            findingAidUrl: null,
-            id: "34275ed5-7123-4104-87ad-c0d9b7927b72",
-            keywords: [""],
-            name: "First List item",
-            published: false,
-            works: [],
-            representativeImage: "",
-          },
-        ],
-      },
-    },
-  },
-];
+import { getCollectionsMock } from "../../components/Collection/collection.gql.mock";
+const mocks = [getCollectionsMock];
 
 function setupTests() {
   return renderWithRouterApollo(
@@ -61,7 +36,7 @@ describe("ScreensCollectionList component", () => {
   it("renders collection list row item from mock", async () => {
     const { getByText } = setupTests();
     await waitFor(() => {
-      expect(getByText("First List item")).toBeInTheDocument();
+      expect(getByText("Great collection")).toBeInTheDocument();
     });
   });
 });
