+++ assets/js/screens/Project/Project.test.js
@@ -1,66 +1,15 @@
 import React from "react";
 import ScreensProject from "./Project";
-import {
-  GET_PROJECT,
-  INGEST_SHEET_STATUS_UPDATES_FOR_PROJECT_SUBSCRIPTION,
-} from "../../components/Project/project.gql.js";
 import { renderWithRouterApollo } from "../../services/testing-helpers";
 import "@testing-library/jest-dom/extend-expect";
 import { Route } from "react-router-dom";
-import { waitFor } from "@testing-library/react";
+import {
+  getProjectMock,
+  ingestSheetUpdatesMock,
+} from "../../components/Project/project.gql.mock";
 
 const MOCK_PROJECT_TITLE = "Mock project title";
-const mocks = [
-  {
-    request: {
-      query: GET_PROJECT,
-      variables: {
-        projectId: "01DNFK4B8XASXNKBSAKQ6YVNF3",
-      },
-    },
-    result: {
-      data: {
-        project: {
-          id: "01DNFK4B8XASXNKBSAKQ6YVNF3",
-          title: MOCK_PROJECT_TITLE,
-          ingestSheets: [
-            {
-              id: "01DNFK56MEN9H0C4CDBE7TECJT",
-              name: "fffff",
-              status: "UPLOADED",
-              updatedAt: "2019-10-07T16:16:57",
-            },
-            {
-              id: "01DNFK9XNJ1FWE8GQGSTR3D1NE",
-              name: "not a csv",
-              status: "COMPLETED",
-              updatedAt: "2019-10-07T16:16:57",
-            },
-          ],
-          updatedAt: "2019-10-07T16:16:57",
-        },
-      },
-    },
-  },
-  {
-    request: {
-      query: INGEST_SHEET_STATUS_UPDATES_FOR_PROJECT_SUBSCRIPTION,
-      variables: {
-        projectId: "01DNFK4B8XASXNKBSAKQ6YVNF3",
-      },
-    },
-    result: {
-      data: {
-        ingestSheetUpdatesForProject: {
-          id: "01DNFK56MEN9H0C4CDBE7TECJT",
-          name: "fffff",
-          status: "VALID",
-          updatedAt: "2019-10-07T16:16:57",
-        },
-      },
-    },
-  },
-];
+const mocks = [getProjectMock, ingestSheetUpdatesMock];
 
 function setupMatchTests() {
   return renderWithRouterApollo(
