+++ assets/js/services/testing-helpers.js
@@ -88,139 +88,3 @@ export function renderWithApollo(ui, { mocks = [] }) {
     ...render(ui, { wrapper: Wrapper }),
   };
 }
-
-export const mockWork = {
-  accessionNumber: "Example-34",
-  administrativeMetadata: {
-    preservationLevel: {
-      id: "0",
-      label: "Level 0",
-    },
-    status: {
-      id: "DONE",
-      label: "Done",
-    },
-  },
-  collection: {
-    id: "147238",
-    name: "Collection Name",
-  },
-  fileSets: [
-    {
-      accessionNumber: "Example-34-3",
-      role: "AM",
-      id: "01DV4BAEAGKNT5P3GH10X263K1",
-      metadata: {
-        description: "Lorem Ipsum",
-        original_filename: "foo.tiff",
-        location: "s3://bucket/foo/bar",
-        label: "foo.tiff",
-        sha256: "123481439834-098349-8",
-      },
-    },
-    {
-      accessionNumber: "Example-34-4",
-      id: "01DV4BAEANHGYQKQ2EPBWJVJSR",
-      role: "AM",
-      metadata: {
-        description: "Lorem Ipsum",
-        original_filename: "foo.tiff",
-        location: "s3://bucket/foo/bar",
-        label: "foo.tiff",
-        sha256: "123481439834-098349-8",
-      },
-    },
-  ],
-  id: "01DV4BAE9NDQHSMRHKM8KC4FNC",
-  insertedAt: "2019-12-02T22:22:30",
-  descriptiveMetadata: {
-    title: "Ima title",
-    description: "Ima description",
-    contributor: [
-      {
-        id: "http://id.gov.loc/someone",
-        label: "Some One",
-        role: {
-          id: "aut",
-          label: "Author",
-          scheme: "MARC_RELATOR",
-        },
-      },
-    ],
-    creator: [
-      {
-        id: "https://theurioftheresource",
-        label: "This is the label",
-      },
-    ],
-    genre: [
-      {
-        id: "https://theurioftheresource",
-        label: "This is the label",
-      },
-    ],
-    language: [
-      {
-        id: "https://theurioftheresource",
-        label: "This is the label",
-      },
-    ],
-    license: {
-      id: "https://theurioftheresource",
-      label: "This is the label",
-    },
-    location: [
-      {
-        id: "https://theurioftheresource",
-        label: "This is the label",
-      },
-    ],
-    rightsStatement: {
-      id: "https://theurioftheresource",
-      label: "This is the label",
-    },
-    stylePeriod: [
-      {
-        id: "https://theurioftheresource",
-        label: "This is the label",
-      },
-    ],
-    subject: [
-      {
-        id: "http://id.gov.loc/something",
-        label: "Some Thing",
-        role: {
-          id: "topical",
-          label: "Topical",
-          scheme: "SUBJECT",
-        },
-      },
-    ],
-    technique: [
-      {
-        id: "https://theurioftheresource",
-        label: "This is the label",
-      },
-    ],
-  },
-  updatedAt: "2019-12-02T22:22:30",
-  published: false,
-  project: {
-    id: "28b6dd45-ef3e-45df-b380-985c9af8b495",
-    name: "Foo",
-  },
-  sheet: {
-    id: "28b6dd45-ef3e-45df-b380-985c9af8b495",
-    name: "Bar",
-  },
-  manifestUrl: "http://foobar",
-  representativeImage: "http://foobar",
-  visibility: {
-    id: "OPEN",
-    label: "Public",
-  },
-  workType: {
-    id: "IMAGE",
-    label: "Image",
-  },
-};
